+++ harfbuzz/src/hb-face.h
@@ -121,25 +121,6 @@ hb_face_get_table_tags (const hb_face_t *face,
 			unsigned int *table_count, /* IN/OUT */
 			hb_tag_t     *table_tags /* OUT */);
 
-
-/*
- * Character set.
- */
-
-HB_EXTERN void
-hb_face_collect_unicodes (hb_face_t *face,
-			  hb_set_t  *out);
-
-HB_EXTERN void
-hb_face_collect_variation_selectors (hb_face_t *face,
-				     hb_set_t  *out);
-
-HB_EXTERN void
-hb_face_collect_variation_unicodes (hb_face_t *face,
-				    hb_codepoint_t variation_selector,
-				    hb_set_t  *out);
-
-
 /*
  * Builder face.
  */